Fixes how the responses for collections of elements are handled
When a collection is returned in response to a RESTful request, the response should now only
include a limited number of fields (the @uuid, @classname, and @noun fields) for each element
plus a URI that can be used to access the details for that element of the collection.  Previously,
the full details were returned whenever a collection of size 1 was detected (meaning that the
response would change format depending on how many elements were in the collection).  The
key that determines the behavior now is whether or not the object being returned responds to
the method named "collection".  If so, then a summary of the elements in the collection is
returned (the proper behavior), otherwise the full details of the object (a single element, not
a collection) will be returned.  This is much more consistent behavior than exists even in the
current version of Razor (which changes what it returns depending on whether the collection
that is being returned has a single element in it or multiple elements in it).  To more closely
reflect this change, we have also refactored the method names that are used to return an object
(either a single object or a collection of objects) as part of the request.  That functionality is
now under the 'slice_success_object' method in 'helpers' section of the various API classes
(in previous versions of these classes, this appeared as the 'slice_success_object_array'
method) and it invokes the 'Razor::WebService::Utils::rz_slice_success_object' (rather than the
'Razor::WebService::Utils::rz_slice_success_object_array' method that was invoked by previous
versions of this same method)