Removes references to Razor-Microkernel
The previous version of this file included references to the
Puppet Labs Razor-Microkernel project (indirectly, via the
link that was provided to download images through the
Puppet Labs download site). This commit removes that
(outdated) reference and replaces it with details on where to
obtain a Hanlon-Microkernel image for use with Hanlon and a
brief overview of the types of Microkernel images that are
available, which provides Hanlon users with a description
of the differences between them.