Added code to sort config output
With these changes, the configuration is output by the config slice in a key-sorted order (with the "@noun" key removed).  This makes it much simpler to find the configuration parameter you are interested in; prior to this set of changes, the configuration parameters were output by the config slice in a semi-random order (based on how they were inserted into the hash-map used to store them) making it very difficult to find a specific configuration parameter.