Fixes #35 (handle missing Hanlon server properly)
The changes in this commit ensure that an understandable error
is output if the CLI-based client is run while the Hanlon
server is not running (or is inaccessible at the URL that the
client is attempting to connect to based on its configuration).
This commit also includes a small fix to avoid "null" values
being output in the help messages returned by the config slice
when these new errors are thrown.