Fixes #107 ("Hanlon subnet" check fails under Tomcat)
When Hanlon is deployed under Tomcat, calls to `Facter.value(:kernel)`
return an empty string. As a result, attempts to get the list of available
interfaces, a value from a specific interface, or the output containing
the text from a single interface using Facter all return either nil or an
empty string. This commit monkey-patches the Facter::Util::IP module
(within the Hanlon::WebService::Utils module) so that the correct
values are returned when the `Facter::Util::IP.get_interfaces` or
`Facter::Util::IP.get_interface_value(interface, label)` methods
are called (even when Hanlon is deployed under Tomcat)