Fixes #195 (Hanlon client configuration isn't returning the local configuration)
The changes in this commit ensure that the `hanlon config` command returns the (local)
client configuration and the `hanlon config server` returns the server configuration. This
is done by modifying the `hanlon config` command so that it uses the local codebase
to access the client configuration from the local filesystem (rather than attempting to
access it from the RESTful API of the server). Since the RESTful API is no longer used
to obtain the (local) client configuration, we can go back to having two endpoints in the
RESTful API of the `/config` resource, the `/config` endpoint itself (which supports
a single `GET` operation that returns the server configuration) and the `/config/ipxe`
endpoint (which returns the iPXE configuration that should be used with the Hanlon
server in response to a `GET /config/ipxe`)