Fixes #208 (extraneous tab in RHEL7 os_boot.erb)
This (minor) commit fixes an issue with the code in this template file that is used
to update the /etc/hosts file with a line that maps the routable IP address for the
node to the hostname and domainname from the active_model instance that was
bound to that node. Previous versions included a '\t' string that was breaking this
mapping, this new version actually outputs a tab instead of the '\t' string.