Fixes #228 (Adds ability to remove active_model bound to a specific node)
Adds ability to remove a specific active_model instance, based on the hardware_id (SMBIOS UUID)
or the Hanlon-assigned UUID of the node  that active_model instance is bound to. Also adds the
ability to get a specific active_model instance based on those same two values (either the
hardware_id or the Hanlon-assigned UUID of the node that active_model instance is bound
to...previously only a search by hardware_id of the bound node was supported).