Initial spec file for api_image_v1.rb. All endpoints tested for basic functionality.
This test requires an ISO file for testing the creation of an image.  To accomplish this, we look for a Hanlon MK image in the ./test/res folder and if not found, we download it from GitHub.  The ISO we look for is hard-coded (bleh) so it would be good to provide a better way.

The ./test/res/README file identifies the function of the (res)source folder.