Updating Gemfile.lock for new gems
With the updated Gemfile, 'bundle udpate' will update the Gemfile.lock file to reflect specific gem versions which should be installed when installing on a vanilla system.  These gem versions have been tested in development, and having the Gemfile.lock as part of the project makes sure that newer gems (which may be incompatible) are not installed on vanilla systems.